+++ src/main/java/org/bytedeco/javacv/Frame.java
@@ -61,6 +61,9 @@ public class Frame implements AutoCloseable, Indexable {
     /** A flag set by a FrameGrabber or a FrameRecorder to indicate a key frame. */
     public boolean keyFrame;
 
+    /** The type of the image frame ('I', 'P', 'B', etc). */
+    public char pictType;
+
     /** Constants to be used for {@link #imageDepth}. */
     public static final int
             DEPTH_BYTE   =  -8,
@@ -125,6 +128,7 @@ public class Frame implements AutoCloseable, Indexable {
         this.imageDepth = depth;
         this.imageChannels = channels;
         this.imageStride = imageStride;
+        this.pictType = '\0';
         this.image = new Buffer[1];
         this.data = null;
         this.streamIndex = -1;
@@ -216,6 +220,7 @@ public class Frame implements AutoCloseable, Indexable {
         newFrame.imageChannels = imageChannels;
         newFrame.imageStride = imageStride;
         newFrame.keyFrame = keyFrame;
+        newFrame.pictType = pictType;
         newFrame.streamIndex = streamIndex;
         newFrame.opaque = new Pointer[3];
         if (image != null) {